+++ public/stylesheets/index.css
@@ -75,7 +75,7 @@ main {
 }
 
 button:active {
-    scale: 0.8;
+    scale: 0.9;
 }
 
 input {
